Refine sidebar rail disclosure animation
